+++ otreeutils/pages.py
@@ -1,3 +1,10 @@
+"""
+oTree page extensions.
+
+Sept. 2018, Markus Konrad <markus.konrad@wzb.eu>
+"""
+
+
 import settings
 from distutils.version import StrictVersion