+++ CHANGES.md
@@ -1,5 +1,12 @@
 # Changes
 
+## v0.4.0 – 2018-09-27
+
+* add admin extensions:
+  * live session data with custom data models
+  * app data export with custom data models in CSV, Excel and JSON formats
+
+
 ## v0.3.0 – 2018-04-25
 
 * made compatible with oTree v2.0